src/functions/requirements.txt
azure-functions==1.21.0
azure-storage-blob==12.19.0
azure-cosmos==4.6.0
pyodbc==5.1.0
